backend/gpu_inference/pipeline.py
"""
AI Virtual Try-On Pipeline
Orchestrates all AI models for realistic garment try-on
"""
import torch
import cv2
import numpy as np
from PIL import Image
from typing import Tuple
import time


class VirtualTryonPipeline:
    """
    Complete AI pipeline for virtual try-on
    
    Pipeline stages:
    1. Human Parsing (SCHP) - Segment body parts
    2. Pose Estimation (OpenPose/MediaPipe) - Extract keypoints
    3. Garment Warping (TPS) - Align garment to pose
    4. Diffusion Try-On (IDM-VTON) - Generate realistic composite
    5. Refinement (RealESRGAN) - Upscale to HD
    """
    
    def __init__(self, device="cuda"):
        """
        Initialize AI pipeline
        
        Args:
            device: torch device ("cuda" or "cpu")
        """
        self.device = device
        print(f"🔧 Initializing AI pipeline on {device}...")
        
        # Load models (lazy loading in production)
        self.schp_model = None
        self.pose_model = None
        self.tryon_model = None
        self.upscaler_model = None
        
        print("✅ AI pipeline initialized")
    
    def load_models(self):
        """Load all AI models into memory"""
        if self.schp_model is None:
            print("📦 Loading SCHP (human parsing)...")
            # self.schp_model = self._load_schp()
            # Placeholder - actual implementation would load from checkpoint
            print("✅ SCHP loaded")
        
        if self.pose_model is None:
            print("📦 Loading Pose Estimation...")
            # self.pose_model = self._load_pose_estimator()
            # Placeholder - actual implementation
            print("✅ Pose model loaded")
        
        if self.tryon_model is None:
            print("📦 Loading IDM-VTON (diffusion try-on)...")
            # self.tryon_model = self._load_idm_vton()
            # Placeholder - actual implementation
            print("✅ IDM-VTON loaded")
        
        if self.upscaler_model is None:
            print("📦 Loading RealESRGAN (upscaler)...")
            # self.upscaler_model = self._load_realesrgan()
            # Placeholder - actual implementation
            print("✅ RealESRGAN loaded")
    
    def _load_schp(self):
        """Load SCHP model for human parsing"""
        # TODO: Implement actual SCHP loading
        # from models.schp import SCHP
        # model = SCHP().to(self.device)
        # model.load_state_dict(torch.load('weights/schp.pth'))
        # model.eval()
        # return model
        pass
    
    def _load_pose_estimator(self):
        """Load pose estimation model"""
        # TODO: Implement actual pose model loading
        # Option 1: MediaPipe (lightweight)
        # Option 2: OpenPose (more accurate)
        pass
    
    def _load_idm_vton(self):
        """Load IDM-VTON diffusion model"""
        # TODO: Implement actual IDM-VTON loading
        # from diffusers import StableDiffusionPipeline
        # pipeline = StableDiffusionPipeline.from_pretrained(...)
        # return pipeline
        pass
    
    def _load_realesrgan(self):
        """Load RealESRGAN upscaler"""
        # TODO: Implement actual RealESRGAN loading
        # from realesrgan import RealESRGAN
        # model = RealESRGAN(device=self.device)
        # return model
        pass
    
    def preprocess_image(self, image_path: str, size: Tuple[int, int] = (512, 768)) -> np.ndarray:
        """
        Load and preprocess image
        
        Args:
            image_path: Path to image file
            size: Target size (width, height)
        
        Returns:
            Preprocessed image as numpy array
        """
        img = Image.open(image_path).convert('RGB')
        img = img.resize(size, Image.LANCZOS)
        img_array = np.array(img)
        return img_array
    
    def parse_human(self, image: np.ndarray) -> np.ndarray:
        """
        Stage 1: Human parsing - segment body parts
        
        Args:
            image: Input image (512x768)
        
        Returns:
            Segmentation mask (512x768)
        """
        print("  📍 Stage 1: Human parsing...")
        start_time = time.time()
        
        # TODO: Implement actual SCHP inference
        # with torch.no_grad():
        #     mask = self.schp_model.predict(image)
        
        # Placeholder: return dummy mask
        mask = np.zeros((image.shape[0], image.shape[1]), dtype=np.uint8)
        
        elapsed = (time.time() - start_time) * 1000
        print(f"  ✅ Parsing complete ({elapsed:.0f}ms)")
        
        return mask
    
    def estimate_pose(self, image: np.ndarray) -> np.ndarray:
        """
        Stage 2: Pose estimation - extract keypoints
        
        Args:
            image: Input image (512x768)
        
        Returns:
            Pose keypoints (18 x 3) - [x, y, confidence]
        """
        print("  📍 Stage 2: Pose estimation...")
        start_time = time.time()
        
        # TODO: Implement actual pose estimation
        # keypoints = self.pose_model.detect(image)
        
        # Placeholder: return dummy keypoints
        keypoints = np.zeros((18, 3), dtype=np.float32)
        
        elapsed = (time.time() - start_time) * 1000
        print(f"  ✅ Pose detected ({elapsed:.0f}ms)")
        
        return keypoints
    
    def warp_garment(
        self,
        garment: np.ndarray,
        keypoints: np.ndarray,
        mask: np.ndarray
    ) -> np.ndarray:
        """
        Stage 3: Garment warping using TPS
        
        Args:
            garment: Garment image (512x768)
            keypoints: Pose keypoints
            mask: Body segmentation mask
        
        Returns:
            Warped garment aligned to pose
        """
        print("  📍 Stage 3: Garment warping...")
        start_time = time.time()
        
        # TODO: Implement TPS warping
        # warped = apply_tps_transform(garment, keypoints, mask)
        
        # Placeholder: return garment as-is
        warped = garment.copy()
        
        elapsed = (time.time() - start_time) * 1000
        print(f"  ✅ Warping complete ({elapsed:.0f}ms)")
        
        return warped
    
    def diffusion_tryon(
        self,
        person_img: np.ndarray,
        garment_img: np.ndarray,
        pose: np.ndarray,
        mask: np.ndarray,
        num_steps: int = 20
    ) -> np.ndarray:
        """
        Stage 4: Diffusion-based virtual try-on (IDM-VTON)
        
        This is the core AI magic - generates photorealistic try-on
        
        Args:
            person_img: Person image (512x768)
            garment_img: Warped garment image
            pose: Pose keypoints
            mask: Body mask
            num_steps: Number of diffusion steps (more = better quality)
        
        Returns:
            Try-on result (512x768)
        """
        print(f"  📍 Stage 4: Diffusion try-on ({num_steps} steps)...")
        start_time = time.time()
        
        # TODO: Implement actual IDM-VTON inference
        # result = self.tryon_model(
        #     image=person_img,
        #     garment=garment_img,
        #     pose=pose,
        #     mask=mask,
        #     num_inference_steps=num_steps,
        #     guidance_scale=7.5
        # ).images[0]
        
        # Placeholder: simple blend
        result = cv2.addWeighted(person_img, 0.5, garment_img, 0.5, 0)
        
        elapsed = (time.time() - start_time) * 1000
        print(f"  ✅ Diffusion complete ({elapsed:.0f}ms)")
        
        return result
    
    def upscale_result(self, image: np.ndarray, scale: int = 2) -> np.ndarray:
        """
        Stage 5: Upscale result to HD using RealESRGAN
        
        Args:
            image: Try-on result (512x768)
            scale: Upscaling factor (2x = 1024x1536)
        
        Returns:
            HD upscaled image
        """
        print(f"  📍 Stage 5: Upscaling ({scale}x)...")
        start_time = time.time()
        
        # TODO: Implement actual RealESRGAN inference
        # upscaled = self.upscaler_model.enhance(image, outscale=scale)
        
        # Placeholder: simple resize
        h, w = image.shape[:2]
        upscaled = cv2.resize(image, (w * scale, h * scale), interpolation=cv2.INTER_CUBIC)
        
        elapsed = (time.time() - start_time) * 1000
        print(f"  ✅ Upscaling complete ({elapsed:.0f}ms)")
        
        return upscaled
    
    @torch.inference_mode()
    def run(
        self,
        person_img_path: str,
        garment_img_path: str,
        num_diffusion_steps: int = 20
    ) -> np.ndarray:
        """
        Run complete virtual try-on pipeline
        
        Args:
            person_img_path: Path to person image
            garment_img_path: Path to garment image
            num_diffusion_steps: Diffusion quality (20=fast, 30=balanced, 50=best)
        
        Returns:
            Final try-on result as numpy array
        """
        print("\n🎨 Starting virtual try-on pipeline...")
        pipeline_start = time.time()
        
        # Load models if not already loaded
        self.load_models()
        
        # Preprocess images
        print("📥 Loading images...")
        person_img = self.preprocess_image(person_img_path, size=(512, 768))
        garment_img = self.preprocess_image(garment_img_path, size=(512, 768))
        
        # Stage 1: Human parsing
        mask = self.parse_human(person_img)
        
        # Stage 2: Pose estimation
        pose = self.estimate_pose(person_img)
        
        # Stage 3: Garment warping
        warped_garment = self.warp_garment(garment_img, pose, mask)
        
        # Stage 4: Diffusion try-on
        tryon_result = self.diffusion_tryon(
            person_img,
            warped_garment,
            pose,
            mask,
            num_steps=num_diffusion_steps
        )
        
        # Stage 5: Upscale to HD
        final_result = self.upscale_result(tryon_result, scale=2)
        
        # Calculate total time
        total_time = (time.time() - pipeline_start) * 1000
        print(f"\n✨ Pipeline complete! Total time: {total_time:.0f}ms ({total_time/1000:.1f}s)")
        
        return final_result
    
    def save_result(self, image: np.ndarray, output_path: str):
        """Save result image to file"""
        img_pil = Image.fromarray(image)
        img_pil.save(output_path, quality=95)
        print(f"💾 Result saved to: {output_path}")


# Example usage
if __name__ == "__main__":
    # Initialize pipeline
    pipeline = VirtualTryonPipeline(device="cuda" if torch.cuda.is_available() else "cpu")
    
    # Run try-on
    result = pipeline.run(
        person_img_path="person.jpg",
        garment_img_path="garment.jpg",
        num_diffusion_steps=20
    )
    
    # Save result
    pipeline.save_result(result, "result.jpg")
