backend/gpu_inference/requirements.txt
# PyTorch (install separately with CUDA)
# torch==2.0.1+cu118
# torchvision==0.15.2+cu118

# Diffusion Models
diffusers==0.25.0
transformers==4.36.2
accelerate==0.25.0

# Computer Vision
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
scikit-image==0.22.0
pillow==10.2.0

# Pose Estimation
mediapipe==0.10.9

# Image Processing
imageio==2.33.1
scipy==1.11.4
numpy==1.26.3

# Utilities
redis==5.0.1
boto3==1.34.23
python-dotenv==1.0.0
tqdm==4.66.1

# Inference Optimization
onnxruntime-gpu==1.16.3
torch-tensorrt==1.4.0
